requirements.txt
boto3==1.20.20
psycopg2-binary==2.9.9
botocore==1.23.54
pendulum==3.1.0
python-dateutil==2.9.0.post0
s3transfer==0.5.2
six==1.17.0
tzdata==2025.2
urllib3==1.26.20
requests==2.31.0
schedule==1.2.0
base58==2.1.1
tronpy==0.6.1
